[JS] fix some imports after refactoring
