+++ lib/js_service_encointer/src/service/encointer.js
@@ -2,7 +2,7 @@ import { assert, bnToU8a, hexToU8a } from '@polkadot/util';
 import { cryptoWaitReady } from '@polkadot/util-crypto';
 import { createType } from '@polkadot/types';
 import { parseEncointerBalance, stringToDegree } from '@encointer/types';
-import { keyring, sendTxWithPair } from './account.js';
+import { keyring } from './keyrings.js';
 import { pallets } from '../config/consts.js';
 import { unsubscribe } from '../utils/unsubscribe.js';
 import { communityIdentifierToString } from '@encointer/util';
@@ -17,6 +17,7 @@ import {
 import { locationFromJson } from '../utils/location-from-json.js';
 import { getFinalizedHeader } from './chain.js';
 import { applyDemurrage } from '../utils/apply-demurrage.js';
+import { sendTxWithPair } from './extrinsics.js';
 
 export async function getCurrentPhase () {
   return api.query.encointerScheduler.currentPhase();
