+++ lib/js_service_encointer/src/service/settings.js
@@ -4,7 +4,7 @@ import { EncointerWorker } from '@encointer/worker-api';
 import { options } from '@encointer/node-api';
 import WS from 'websocket';
 import { pallets } from '../config/consts.js';
-import { keyring } from './account.js';
+import { keyring } from './keyrings.js';
 
 const { w3cwebsocket: WebSocket } = WS;
